# docs/es/propuesta_motor/app/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para el Proof of Concept del Motor Contable Colombiano",
        "title": "Motor Contable POC API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Soporte API",
            "email": "soporte@motorcontable.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/vouchers": {
            "get": {
                "description": "Retorna una lista paginada de comprobantes de la organización",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comprobantes"
                ],
                "summary": "Lista de comprobantes",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Número de página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Elementos por página",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de comprobantes",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.VouchersListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Parámetros inválidos",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Crea un nuevo comprobante contable con sus líneas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comprobantes"
                ],
                "summary": "Crear comprobante",
                "parameters": [
                    {
                        "description": "Datos del comprobante",
                        "name": "voucher",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VoucherCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Comprobante creado",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Voucher"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vouchers/by-date-range": {
            "get": {
                "description": "Retorna comprobantes filtrados por rango de fechas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comprobantes"
                ],
                "summary": "Comprobantes por fechas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fecha inicial (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Fecha final (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Número de página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Elementos por página",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de comprobantes",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.VouchersListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Parámetros inválidos",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vouchers/generate": {
            "post": {
                "description": "Genera un comprobante automáticamente usando una plantilla DSL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comprobantes"
                ],
                "summary": "Generar desde plantilla",
                "parameters": [
                    {
                        "description": "Datos para generación",
                        "name": "generation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Comprobante generado",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Voucher"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vouchers/{id}": {
            "get": {
                "description": "Retorna el detalle completo de un comprobante específico",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comprobantes"
                ],
                "summary": "Detalle de comprobante",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del comprobante",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del comprobante",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.VoucherDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Comprobante no encontrado",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vouchers/{id}/cancel": {
            "post": {
                "description": "Cambia el estado del comprobante a cancelado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comprobantes"
                ],
                "summary": "Cancelar comprobante",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del comprobante",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comprobante cancelado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "No se puede cancelar",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    },
                    "404": {
                        "description": "Comprobante no encontrado",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vouchers/{id}/post": {
            "post": {
                "description": "Cambia el estado del comprobante a contabilizado y genera asiento",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comprobantes"
                ],
                "summary": "Contabilizar comprobante",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del comprobante",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comprobante contabilizado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "No se puede contabilizar",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    },
                    "404": {
                        "description": "Comprobante no encontrado",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.PaginationInfo": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "pages": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "models.StandardResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/models.ErrorInfo"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "models.Voucher": {
            "type": "object",
            "properties": {
                "cost_center_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_user_id": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_balanced": {
                    "type": "boolean"
                },
                "number": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "string"
                },
                "period_id": {
                    "type": "string"
                },
                "posted_at": {
                    "type": "string"
                },
                "posted_by_user_id": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                },
                "status": {
                    "description": "DRAFT, POSTED, CANCELLED",
                    "type": "string"
                },
                "third_party_id": {
                    "type": "string"
                },
                "total_credit": {
                    "type": "number"
                },
                "total_debit": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "voucher_lines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.VoucherLine"
                    }
                },
                "voucher_type": {
                    "description": "JOURNAL, PURCHASE, SALE, PAYMENT, RECEIPT, ADJUSTMENT",
                    "type": "string"
                }
            }
        },
...
        "models.VoucherCreateRequest": {
            "type": "object",
            "required": [
                "date",
                "description",
                "voucher_lines",
                "voucher_type"
            ],
            "properties": {
                "additional_data": {
                    "$ref": "#/definitions/models.VoucherAdditionalData"
                },
                "cost_center_id": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                },
                "third_party_id": {
                    "type": "string"
                },
                "voucher_lines": {
                    "type": "array",
                    "minItems": 2,
                    "items": {
                        "$ref": "#/definitions/models.VoucherLineRequest"
                    }
                },
                "voucher_type": {
                    "type": "string"
                }
            }
        },
...
        "models.VouchersListResponse": {
            "type": "object",
            "properties": {
                "pagination": {
                    "$ref": "#/definitions/models.PaginationInfo"
                },
                "vouchers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Voucher"
                    }
                }
            }
        }
    }
}